fixed 693

+++ cvxpy/problems/problem.py
@@ -31,6 +31,7 @@ from cvxpy.constraints import Equality, Inequality, NonPos, Zero
 import cvxpy.utilities as u
 from collections import namedtuple
 import multiprocess as multiprocessing
+import numpy as np
 
 
 SolveResult = namedtuple(
@@ -779,7 +780,7 @@ class SizeMetrics(object):
                 self.max_data_dimension = big
 
             if self.max_big_small_squared < big*small**2:
-                self.max_big_small_squared = big*small**2
+                self.max_big_small_squared = np.int64(big)*np.int64(small)**2
 
         # num_scalar_eq_constr
         self.num_scalar_eq_constr = 0
